src/pyproject.toml
[tool.poetry]
authors = ["Your Name <your_email@example.com>"]
description = "A short description of your project."
name = "your_project_name"
version = "0.1.0"

[tool.poetry.dependencies]
celery = "^5.3"
celery-redbeat = "^2.2"
cryptography = "^42.0"
defusedxml = "^0.7"
django = "^5.1"
django-anymail = "^10.2"
django-cors-headers = "^4.4"
django-extensions = "^3.2"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
gunicorn = "^21.2"
oauthlib = "^3.2"
pillow = "^10.2"
psycopg2 = "^2.9"
pydantic = "^2.5"
pyjwt = "^2.8"
pyopenssl = "^24.0"
python = "^3.10"
python-dateutil = "^2.8"
python-decouple = "^3.8"
python-dotenv = "^1.0"
python-jose = "^3.3"
redis = "^5.0"
requests = "^2.31"
sentry-sdk = "^1.39"
shortuuid = "^1.0"
social-auth-app-django = "^5.4"
social-auth-core = "^4.5"
uvicorn = {version = "^0.27", extras = ["standard"]}
watchgod = "^0.8"
werkzeug = "^3.0"
whitenoise = "^6.6"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2"
freezegun = "^1.4"
markdown-urlize = "^0.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6"
mkdocs-awesome-pages-plugin = "^2.9.2"
mkdocs-graphviz = "^1.5.3"
mkdocs-material = "^9.5.6"
mkdocstrings = "^0.26"
mkdocstrings-python = "^1.11"
pydot = "^3.0.0"
# Additional development tools
pytest = "^7.4"
ruff = "^0.6.7"
[tool.ruff]
line-length = 120
target-version = "py312"
# Exclude a variety of commonly ignored directories.
exclude = [
  "migrations",
  "asgi.py",
]
[tool.ruff.lint]
extend-select = ["I"]
ignore = [
  "ANN",
  "COM812",
  "ISC001",
  "D",
  "PT009",
  "B904",
  "S101",
  "ARG001",
  "ARG002",
  "G004",
  "EM102",
  "TRY003",
  "TRY401",
  "EM101",
  "S311",
  "S107",
  "S105",
  "N802",
  "TCH001",
  "TCH002",
  "PTH",
  "RUF012",
  "ERA001", # temporary
]
select = ["ALL"]

per-file-ignores = {}
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.format]
# Like Black, use double quotes for strings.
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
